HACK: neversleep: A utility to disable deep sleep
Because of a prevalent bug concerning suspending to deep sleep on the
tone platform, introduce a workaround.

Hold an infinite wakelock so the device never enters deep sleep and thus
has no problems waking up in a timely manner.

Because this is draining battery, introduce a property to enable or
disable this behaviour.

Change-Id: I15ba18432ff1f2ccd6f2c9c800e6bbb643d49806