recovery fstab: Add FRP and DSP partitions
The FRP partition is used for the "Factory Reset
Protection" feature introduced in Marshmallow.
Allow wiping it to get around that "feature".

The DSP partition holds firmware files for the
Qualcomm Hexagon DSP. In case of corrupted files,
allow backing up or flashing an image to avoid
having to use desktop tools like Emma or
FlashTool.

Change-Id: I757dc7d2913798a457372e39adb1177e11713edf